refactor(core-agent): move spark-command agent from stage-ui to core-agent
- Relocate spark-command and spark-notify agents to @proj-airi/core-agent for
  better separation of concerns and reusability across packages
- Add barrel export and package.json entry for ./agents/spark-command
- Merge duplicate sparkNotifyCommandGuidanceSchema with sparkCommandGuidanceSchema
- Remove sparkNotifyCommandItemSchema (consolidated into sparkCommandToolSchema)
- Add JSDoc with @example blocks to normalize functions for documentation
- Update provider test imports to use new @proj-airi/core-agent path
- Update tool-resolver to import from core-agent instead of local tools
- Delete legacy re-export shims in stage-ui/src/tools/character
