+++ packages/core-agent/src/agents/spark-command/index.ts
@@ -1,0 +1,17 @@
+export {
+  normalizeSparkCommandDestinations,
+  normalizeSparkCommandGuidanceOptions,
+  normalizeSparkCommandMetadata,
+  normalizeSparkCommandPersona,
+  normalizeSparkCommandStringList,
+  normalizeSparkCommandStringValue,
+  sparkCommandGuidanceOptionSchema,
+  sparkCommandGuidanceSchema,
+  sparkCommandIntentSchema,
+  sparkCommandInterruptSchema,
+  sparkCommandPersonaSchema,
+  sparkCommandPrioritySchema,
+  sparkCommandToolSchema,
+} from './schema'
+export type { CreateSparkCommandToolOptions } from './tools'
+export { createSparkCommandTool } from './tools'
