+++ packages/core-agent/tsdown.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from 'tsdown'
 export default defineConfig({
   entry: [
     'src/index.ts',
+    'src/agents/spark-command/index.ts',
     'src/agents/spark-notify/index.ts',
   ],
   dts: true,
