+++ packages/stage-ui/src/libs/providers/providers/openrouter-ai/index.test.ts
@@ -2,10 +2,9 @@ import type { ChatRequestOptions } from '@proj-airi/provider-inference'
 import type { ChatProviderWithExtraOptions } from '@xsai-ext/providers/utils'
 import type { JsonSchema } from 'xsschema'
 
+import { createSparkCommandTool } from '@proj-airi/core-agent/agents/spark-command'
 import { getDefinedProvider } from '@proj-airi/provider-inference'
 import { afterEach, describe, expect, it, vi } from 'vitest'
-
-import { createSparkCommandTool } from '../../../../tools/character/orchestrator/spark-command'
 
 interface ChatRequestBody {
   tools: Array<{
