--- packages/stage-ui/src/tools/character/index.ts
@@ -1,1 +1,0 @@
-export * from './orchestrator'
