--- packages/stage-ui/src/tools/character/orchestrator/index.ts
@@ -1,3 +1,0 @@
-export * from './spark-command'
-export * from './spark-command-shared'
-export * from './spark-notify'
