--- packages/stage-ui/src/tools/character/orchestrator/spark-notify.ts
@@ -1,9 +1,0 @@
-export type {
-  CreateSparkNotifyToolsOptions,
-  SparkNotifyCommandDraft,
-} from '@proj-airi/core-agent/agents/spark-notify'
-
-export {
-  createSparkNotifyTools,
-  sparkNotifyCommandSchema,
-} from '@proj-airi/core-agent/agents/spark-notify'
