+++ packages/stage-ui/src/tools/index.ts
@@ -1,4 +1,3 @@
-export * from './character'
 export * from './debug'
 export * from './mcp'
 export * from './web-search'
